libft project
This project is redoing a set of functions from the libc. The functions will have the
same prototypes and implement the same behaviors as the originals. All the function comply
with the way they are defined in their man. The only difference will be their names. They
will begin with the ’ft_’ prefix. For instance, strlen becomes ft_strlen.